ple_coupling: fix some error cases at last time step.
When a code has required a last time step previously and another follow
its minimum value, an error could occur at the end of the time step
computation due to a missing "leader", even when that value was not needed
since a computation stop flag was set.

Also cleanup handling of the initialization flag, which was not removed
correctly.